Refactor parsing logic and revise exception handling
